Enhanced and fixed the install script.

+++ install.php
@@ -1,6 +1,8 @@
 #!/usr/bin/env php
-<?php
+<?php declare(strict_types=1);
 
+require_once __DIR__ . '/vendor/autoload.php';
+echo __DIR__ . '/vendor/autoload.php' . "\n";
 /*************************************************
  *      ●●● ●●● ●●● ●●● ●●● ●●● ●●● ●●● ●●●      *
  *     ●●●                               ●●●     *
@@ -39,17 +41,21 @@ function getProjectFiles(): array
 
 function replaceProjectName(string $filename, string $newName): void
 {
+//    dd([$filename, $newName]);
     $origText = file_get_contents($filename);
     $newText = str_replace('Skeleton', $newName, $origText);
-
+//    dd([$origText, $newText]);
     if ($newText !== $origText) {
+//        dd($newText);
         file_put_contents($filename, $newText);
     }
 }
 
 function installGitAttributes()
 {
-    rename('_gitattributes', '.gitattributes');
+    if (file_exists('_gitattributes')) {
+        rename('_gitattributes', '.gitattributes');
+    }
 }
 
 function main()
@@ -62,6 +68,7 @@ function main()
     // Replace "Skeleton" with $project in every PHP file.
     $files = getProjectFiles();
     foreach ($files as $filename) {
+        dump([$filename, $project]);
         replaceProjectName($filename, $project);
     }
 
@@ -73,8 +80,12 @@ function main()
     echo "\nDelete every LICENSE you do not want.\n";
 
     echo "\bNote: This install.php file has deleted itself.\n";
-    unlink('src/.gitkeep');
-    unlink('install.php');
+    if (file_exists('./src/.gitkeey')) {
+        unlink('src/.gitkeep');
+    }
+    if (file_exists('/install.php')) {
+        unlink('install.php');
+    }
 }
 
 main();
